backend/requirements.txt
pydantic>=2.0.0
lxml>=4.9.0
pymupdf>=1.23.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
google-generativeai>=0.4.0
faiss-cpu>=1.7.4
fpdf>=1.7.2
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
feedparser>=6.0.0
tenacity>=8.2.0
cryptography>=41.0.0
PyJWT>=2.8.0
rank-bm25
sentence-transformers
networkx
langchain-huggingface

# API Layer (Sprint 1)
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
httpx>=0.24.0
